Split utils/utility into specific modules
