+++ app/utils/function.js
@@ -1,0 +1,8 @@
+import R from 'ramda';
+
+// argsToArray :: (a, …) -> [a, …]
+const argsToArray = R.unapply(R.identity);
+
+export {
+  argsToArray,
+};
