+++ app/utils/number.js
@@ -1,0 +1,6 @@
+// round :: Number -> Number
+const round = (x) => Math.round(x * 100) / 100;
+
+export {
+  round,
+};
